+++ src/main/java/io/github/astrapi69/collection/iterator/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides adapter classes for iterators
+ */
+package io.github.astrapi69.collection.iterator;