+++ src/main/java/io/github/astrapi69/collection/set/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides utility classes for set objects and set implementations
+ */
+package io.github.astrapi69.collection.set;